Add "continuous" or "character" mode `-c` option
If `termios.h` is available on the system at compile time, support the
`-c` option which will work on input one character at a time. This is
as opposed to buffering the input until the return key is pressed.

Also improve the `keyer_cleanstring()` function to be safer and work
correctly.